refactor: split side-nav and item styles, use styles folder
