+++ packages/side-nav/src/styles/vaadin-side-nav-base-styles.js
@@ -1,0 +1,40 @@
+/**
+ * @license
+ * Copyright (c) 2023 - 2025 Vaadin Ltd.
+ * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
+ */
+import '@vaadin/component-base/src/style-props.js';
+import { css } from 'lit';
+import { sharedStyles } from './vaadin-side-nav-shared-base-styles.js';
+
+const sideNav = css`
+  :host {
+    white-space: nowrap;
+  }
+
+  [part='label'] {
+    align-self: start;
+    display: flex;
+    align-items: center;
+    justify-content: start;
+    gap: var(--vaadin-side-nav-item-gap, var(--_vaadin-gap-container-inline));
+    padding: var(--vaadin-side-nav-item-padding, var(--_vaadin-padding-container));
+    font-size: var(--vaadin-side-nav-label-font-size, 0.875em);
+    font-weight: var(--vaadin-side-nav-label-font-weight, 500);
+    color: var(--vaadin-side-nav-label-color, inherit);
+    line-height: var(--vaadin-side-nav-label-line-height, inherit);
+    border-radius: var(--vaadin-side-nav-item-border-radius, var(--_vaadin-radius-m));
+  }
+`;
+
+export const sideNavStyles = [sharedStyles, sideNav];
+
+export const sideNavSlotStyles = css`
+  vaadin-side-nav:has(vaadin-icon[slot='prefix']) {
+    --_has-prefix-icon: '';
+  }
+
+  vaadin-side-nav-item:has(> vaadin-icon[slot='prefix'])::part(link) {
+    --_has-prefix-icon:;
+  }
+`;
