+++ packages/side-nav/src/styles/vaadin-side-nav-core-styles.d.ts
@@ -1,0 +1,10 @@
+/**
+ * @license
+ * Copyright (c) 2023 - 2025 Vaadin Ltd.
+ * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
+ */
+import type { CSSResult } from 'lit';
+
+export const sideNavStyles: CSSResult;
+
+export const sideNavSlotStyles: CSSResult;
