+++ packages/side-nav/src/styles/vaadin-side-nav-core-styles.js
@@ -1,0 +1,39 @@
+/**
+ * @license
+ * Copyright (c) 2023 - 2025 Vaadin Ltd.
+ * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
+ */
+import { css } from 'lit';
+
+export const sideNavStyles = css`
+  :host {
+    display: block;
+  }
+
+  :host([hidden]) {
+    display: none !important;
+  }
+
+  button {
+    display: flex;
+    align-items: center;
+    justify-content: inherit;
+    width: 100%;
+    margin: 0;
+    padding: 0;
+    background-color: initial;
+    color: inherit;
+    border: initial;
+    outline: none;
+    font: inherit;
+    text-align: inherit;
+  }
+
+  [part='children'] {
+    padding: 0;
+    margin: 0;
+    list-style-type: none;
+  }
+`;
+
+export const sideNavSlotStyles = css``;
