+++ packages/side-nav/src/styles/vaadin-side-nav-item-base-styles.js
@@ -5,126 +5,9 @@
  */
 import '@vaadin/component-base/src/style-props.js';
 import { css } from 'lit';
+import { sharedStyles } from './vaadin-side-nav-shared-base-styles.js';
 
-const sharedStyles = css`
-  :host {
-    display: flex;
-    flex-direction: column;
-    gap: var(--vaadin-side-nav-items-gap, var(--_vaadin-gap-container-block));
-    cursor: default;
-    -webkit-tap-highlight-color: transparent;
-  }
-
-  :host([hidden]),
-  [hidden] {
-    display: none !important;
-  }
-
-  button {
-    appearance: none;
-    margin: 0;
-    padding: 0;
-    border: 0;
-    background: transparent;
-    color: inherit;
-    font: inherit;
-    text-align: inherit;
-    cursor: var(--vaadin-clickable-cursor);
-    flex: none;
-  }
-
-  [part='toggle-button'] {
-    border-radius: var(--vaadin-side-nav-item-border-radius, var(--_vaadin-radius-s));
-    color: var(--_vaadin-color);
-  }
-
-  [part='toggle-button']::before {
-    content: '';
-    display: block;
-    background: currentColor;
-    mask-image: var(--_vaadin-icon-chevron-down);
-    width: var(--vaadin-icon-size, 1lh);
-    height: var(--vaadin-icon-size, 1lh);
-    rotate: -90deg;
-  }
-
-  :host([dir='rtl']) [part='toggle-button']::before {
-    scale: -1;
-  }
-
-  :host(:is(vaadin-side-nav-item[expanded], vaadin-side-nav:not([collapsed]))) [part='toggle-button'] {
-    rotate: 90deg;
-  }
-
-  :host([dir='rtl']:is(vaadin-side-nav-item[expanded], vaadin-side-nav:not([collapsed]))) [part='toggle-button'] {
-    rotate: -90deg;
-  }
-
-  @media (prefers-reduced-motion: no-preference) {
-    [part='toggle-button'] {
-      transition: rotate 150ms;
-    }
-  }
-
-  :host([disabled]) [part='toggle-button'] {
-    opacity: 0.5;
-  }
-
-  [part='children'] {
-    padding: 0;
-    margin: 0;
-    list-style-type: none;
-    display: flex;
-    flex-direction: column;
-    gap: var(--vaadin-side-nav-items-gap, var(--_vaadin-gap-container-block));
-  }
-
-  :focus-visible {
-    outline: var(--vaadin-focus-ring-width) solid var(--vaadin-focus-ring-color);
-  }
-
-  @media (forced-colors: active) {
-    [part='toggle-button']::before {
-      background: CanvasText;
-    }
-  }
-`;
-
-export const sideNavStyles = css`
-  ${sharedStyles}
-
-  :host {
-    white-space: nowrap;
-  }
-
-  [part='label'] {
-    align-self: start;
-    display: flex;
-    align-items: center;
-    justify-content: start;
-    gap: var(--vaadin-side-nav-item-gap, var(--_vaadin-gap-container-inline));
-    padding: var(--vaadin-side-nav-item-padding, var(--_vaadin-padding-container));
-    font-size: var(--vaadin-side-nav-label-font-size, 0.875em);
-    font-weight: var(--vaadin-side-nav-label-font-weight, 500);
-    color: var(--vaadin-side-nav-label-color, inherit);
-    line-height: var(--vaadin-side-nav-label-line-height, inherit);
-    border-radius: var(--vaadin-side-nav-item-border-radius, var(--_vaadin-radius-m));
-  }
-`;
-
-export const sideNavSlotStyles = css`
-  vaadin-side-nav:has(vaadin-icon[slot='prefix']) {
-    --_has-prefix-icon: '';
-  }
-
-  vaadin-side-nav-item:has(> vaadin-icon[slot='prefix'])::part(link) {
-    --_has-prefix-icon:;
-  }
-`;
-
-export const sideNavItemStyles = css`
-  ${sharedStyles}
-
+const sideNavItem = css`
   [part='content'] {
     display: flex;
     align-items: center;
@@ -234,3 +117,5 @@ export const sideNavItemStyles = css`
     }
   }
 `;
+
+export const sideNavItemStyles = [sharedStyles, sideNavItem];
