+++ packages/side-nav/src/styles/vaadin-side-nav-item-core-styles.d.ts
@@ -5,6 +5,4 @@
  */
 import type { CSSResult } from 'lit';
 
-export const sideNavStyles: CSSResult;
-export const sideNavSlotStyles: CSSResult;
 export const sideNavItemStyles: CSSResult;
